Move actual configuration action to its own file.
This to solve circular import issues with units.
