+++ astropy/constants/cgs.py
@@ -3,14 +3,13 @@
 Astronomical and physics constants in cgs units.  See :mod:`astropy.constants`
 for a complete listing of constants defined in Astropy.
 """
-
 import itertools
 
 from .constant import Constant
-from . import codata, iaudata
+from .config import codata, iaudata
 
 for _nm, _c in itertools.chain(sorted(vars(codata).items()),
                                sorted(vars(iaudata).items())):
     if (isinstance(_c, Constant) and _c.abbrev not in locals()
-         and _c.system in ['esu', 'gauss', 'emu']):
+            and _c.system in ['esu', 'gauss', 'emu']):
         locals()[_c.abbrev] = _c
