+++ astropy/constants/si.py
@@ -3,16 +3,13 @@
 Astronomical and physics constants in SI units.  See :mod:`astropy.constants`
 for a complete listing of constants defined in Astropy.
 """
-
-
-
 import itertools
 
 from .constant import Constant
-from . import codata, iaudata
+from .config import codata, iaudata
 
 for _nm, _c in itertools.chain(sorted(vars(codata).items()),
                                sorted(vars(iaudata).items())):
     if (isinstance(_c, Constant) and _c.abbrev not in locals()
-         and _c.system == 'si'):
+            and _c.system == 'si'):
         locals()[_c.abbrev] = _c
